refactor: split OrderFormPage (1231 LOC) into 4 focused components
- OrderStatusActions.tsx — submit/accept/reject/settlement buttons
- OrderCreateForm.tsx — create-mode form with weight calc + map preview
- OrderDetailView.tsx — detail view with actual times, crew, operations, map
- OrderFormPage.tsx — thin orchestrator with hooks + state + dialogs

Zero functional changes — verified by e2e tests (25/25 pass).
