+++ frontend/src/pages/orders/OrderFormPage.tsx
@@ -1,21 +1,20 @@
 /**
- * OrderFormPage — dual create/detail page for flight orders.
+ * OrderFormPage -- dual create/detail orchestrator for flight orders.
  *
- * Create mode:  /orders/new  — helicopter dropdown, crew multi-select with live weight calc,
- *   operation multi-select (status=3 only), landing site dropdowns, safety validation errors.
- * Detail mode:  /orders/:id  — read-only fields, status action buttons, combined map.
+ * Create mode:  /orders/new  -- renders <OrderCreateForm />
+ * Detail mode:  /orders/:id  -- renders <OrderStatusActions /> + <OrderDetailView />
+ *
+ * All state, queries, mutations, and dialogs live here. Child components
+ * receive data and handlers via props.
  */
-import { useState, useEffect, useMemo, type FormEvent } from "react";
+import { useState, useEffect, type FormEvent } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 import { useTranslation } from "react-i18next";
 import { apiFetch, ApiError } from "@/lib/api";
 import { useAuth } from "@/lib/auth";
 import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
 import { Badge } from "@/components/ui/badge";
-import { Select } from "@/components/ui/select";
 import {
   Dialog,
   DialogContent,
@@ -24,9 +23,11 @@ import {
   DialogDescription,
   DialogFooter,
 } from "@/components/ui/dialog";
-import { OrderMap } from "@/components/maps/OrderMap";
-import { ArrowLeft, CheckCircle, XCircle } from "lucide-react";
-import { ORDER_FORM_STATUS_BADGE_CLASS, SYSTEM_ROLE, CREW_ROLE } from "@/lib/constants";
+import { ArrowLeft } from "lucide-react";
+import { ORDER_FORM_STATUS_BADGE_CLASS, SYSTEM_ROLE } from "@/lib/constants";
+import { OrderStatusActions } from "./OrderStatusActions";
+import { OrderCreateForm } from "./OrderCreateForm";
+import { OrderDetailView } from "./OrderDetailView";
 
 // ── Types ──────────────────────────────────────────────────────────
 
@@ -108,13 +109,11 @@ interface OperationDetailForMap {
   route_coordinates: [number, number][] | null;
 }
 
-// Status labels now use t('orders.statusN') via i18n
-
 // ── Helpers ─────────────────────────────────────────────────────────
 
 function toDatetimeLocal(iso: string | null | undefined): string {
   if (!iso) return "";
-  // ISO string → datetime-local input format (YYYY-MM-DDTHH:mm)
+  // ISO string -> datetime-local input format (YYYY-MM-DDTHH:mm)
   return iso.slice(0, 16);
 }
 
@@ -210,33 +209,6 @@ export function OrderFormPage() {
     },
     enabled: isCreate && selectedOpIds.length > 0,
   });
-
-  // Active helicopters only
-  const activeHelicopters = useMemo(
-    () => helicopters.filter((h) => h.status === "aktywny"),
-    [helicopters]
-  );
-
-  // Selected helicopter for weight comparison
-  const selectedHelicopter = useMemo(
-    () => activeHelicopters.find((h) => h.id === Number(helicopterId)),
-    [activeHelicopters, helicopterId]
-  );
-
-  // ── Live weight calculation ────────────────────────────────────
-
-  const selectedCrewMembers = useMemo(
-    () => allCrew.filter((c) => selectedCrewIds.includes(c.id)),
-    [allCrew, selectedCrewIds]
-  );
-
-  const totalCrewWeight = useMemo(() => {
-    const crewWeight = selectedCrewMembers.reduce((sum, c) => sum + c.weight, 0);
-    // Include pilot weight from their crew-member profile (matched by email)
-    const pilotMember = allCrew.find((c) => c.role === CREW_ROLE.PILOT && c.email === user?.email);
-    const pilotWeight = pilotMember?.weight ?? 0;
-    return crewWeight + pilotWeight;
-  }, [selectedCrewMembers, allCrew, user]);
 
   // ── Fetch order for detail mode ────────────────────────────────
   const { data: order, isLoading: loadingOrder } = useQuery<FlightOrderDetail>({
@@ -479,7 +451,7 @@ export function OrderFormPage() {
 
   const currentStatus = order?.status ?? 1;
 
-  // Map data
+  // Map data for detail mode
   const mapOperations = operationDetails
     .filter((op) => op.route_coordinates && op.route_coordinates.length > 0)
     .map((op) => ({
@@ -545,517 +517,80 @@ export function OrderFormPage() {
         </div>
       )}
 
-      {/* Status action buttons (detail mode) */}
-      {!isCreate && (
-        <OrderStatusActions
-          status={currentStatus}
-          isPilot={isPilot}
-          isSupervisor={isSupervisor}
-          hasActualTimes={!!order?.actual_start_datetime && !!order?.actual_end_datetime}
-          onSubmit={() => submitMutation.mutate()}
-          onAccept={() => acceptMutation.mutate()}
-          onReject={() => setShowRejectDialog(true)}
-          onCompletePartial={() => completePartialMutation.mutate()}
-          onCompleteFull={() => completeFullMutation.mutate()}
-          onNotCompleted={() => setShowNotCompletedDialog(true)}
-          submitPending={submitMutation.isPending}
-          acceptPending={acceptMutation.isPending}
-          rejectPending={rejectMutation.isPending}
-          completePartialPending={completePartialMutation.isPending}
-          completeFullPending={completeFullMutation.isPending}
-          notCompletedPending={notCompletedMutation.isPending}
-        />
-      )}
-
+      {/* Render create form OR detail view + status actions */}
       {isCreate ? (
-        /* ═══════════ CREATE MODE ═══════════ */
-        <div className="rounded-md bg-surface-container-low p-6">
-          <form onSubmit={handleSubmit} className="space-y-5">
-            {/* Pilot — auto-filled, read-only */}
-            <div className="space-y-2">
-              <Label>{t('orders.pilotAutoFilled')}</Label>
-              <Input
-                value={
-                  user
-                    ? `${user.first_name} ${user.last_name}`
-                    : ""
-                }
-                disabled
-                className="bg-muted"
-              />
-            </div>
-
-            {/* Planned dates */}
-            <div className="grid grid-cols-2 gap-4">
-              <div className="space-y-2">
-                <Label htmlFor="plannedStart">{t('orders.plannedStartLabel')}</Label>
-                <Input
-                  id="plannedStart"
-                  type="datetime-local"
-                  value={plannedStart}
-                  onChange={(e) => setPlannedStart(e.target.value)}
-                  required
-                />
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="plannedEnd">{t('orders.plannedEndLabel')}</Label>
-                <Input
-                  id="plannedEnd"
-                  type="datetime-local"
-                  value={plannedEnd}
-                  onChange={(e) => setPlannedEnd(e.target.value)}
-                  required
-                />
-              </div>
-            </div>
-
-            {/* Helicopter dropdown */}
-            <div className="space-y-2">
-              <Label htmlFor="helicopter">{t('orders.helicopterLabel')}</Label>
-              {activeHelicopters.length === 0 ? (
-                <p className="text-sm text-muted-foreground">
-                  {t('orders.noActiveHelicopters')}
-                </p>
-              ) : (
-                <Select
-                  id="helicopter"
-                  value={helicopterId}
-                  onChange={(e) => setHelicopterId(e.target.value)}
-                  required
-                >
-                  <option value="">{t('orders.selectHelicopter')}</option>
-                  {activeHelicopters.map((h) => (
-                    <option key={h.id} value={h.id}>
-                      {h.registration_number} — {h.helicopter_type} ({t('orders.helicopterOptionSuffix', { maxPayload: h.max_payload_weight, range: h.range_km })})
-                    </option>
-                  ))}
-                </Select>
-              )}
-            </div>
-
-            {/* Crew multi-select with weight */}
-            <div className="space-y-2">
-              <Label>{t('orders.crewLabel')}</Label>
-              {allCrew.length === 0 ? (
-                <p className="text-sm text-muted-foreground">
-                  {t('orders.noCrewMembers')}
-                </p>
-              ) : (
-                <div className="rounded-md border p-3 max-h-48 overflow-y-auto space-y-1">
-                  {[...allCrew]
-                    .sort((a, b) => `${a.first_name} ${a.last_name}`.localeCompare(`${b.first_name} ${b.last_name}`))
-                    .map((c) => (
-                    <label
-                      key={c.id}
-                      className="flex items-center gap-2 text-sm py-0.5"
-                    >
-                      <input
-                        type="checkbox"
-                        checked={selectedCrewIds.includes(c.id)}
-                        onChange={() => handleCrewToggle(c.id)}
-                        className="rounded border-input"
-                      />
-                      <span>
-                        {c.first_name} {c.last_name}
-                      </span>
-                      <span className="text-muted-foreground">
-                        ({c.weight} kg, {c.role})
-                      </span>
-                    </label>
-                  ))}
-                </div>
-              )}
-              {/* Weight summary */}
-              <div className="text-sm mt-1">
-                <span className="font-medium">{t('orders.crewWeight')}</span>{" "}
-                <span
-                  className={
-                    selectedHelicopter &&
-                    totalCrewWeight > selectedHelicopter.max_payload_weight
-                      ? "text-destructive-foreground font-bold"
-                      : ""
-                  }
-                >
-                  {totalCrewWeight} kg
-                </span>
-                {selectedHelicopter && (
-                  <span className="text-muted-foreground">
-                    {" "}
-                    / {selectedHelicopter.max_payload_weight} kg max
-                    {totalCrewWeight > selectedHelicopter.max_payload_weight && (
-                      <span className="text-destructive-foreground ml-1">
-                        {t('orders.weightExceeded')}
-                      </span>
-                    )}
-                  </span>
-                )}
-              </div>
-            </div>
-
-            {/* Landing sites */}
-            <div className="grid grid-cols-2 gap-4">
-              <div className="space-y-2">
-                <Label htmlFor="startSite">{t('orders.startLandingSite')}</Label>
-                <Select
-                  id="startSite"
-                  value={startSiteId}
-                  onChange={(e) => setStartSiteId(e.target.value)}
-                  required
-                >
-                  <option value="">{t('orders.selectLandingSite')}</option>
-                  {landingSites.map((s) => (
-                    <option key={s.id} value={s.id}>
-                      {s.name}
-                    </option>
-                  ))}
-                </Select>
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="endSite">{t('orders.endLandingSite')}</Label>
-                <Select
-                  id="endSite"
-                  value={endSiteId}
-                  onChange={(e) => setEndSiteId(e.target.value)}
-                  required
-                >
-                  <option value="">{t('orders.selectLandingSite')}</option>
-                  {landingSites.map((s) => (
-                    <option key={s.id} value={s.id}>
-                      {s.name}
-                    </option>
-                  ))}
-                </Select>
-              </div>
-            </div>
-
-            {/* Operations multi-select */}
-            <div className="space-y-2">
-              <Label>{t('orders.operationsConfirmed')}</Label>
-              {confirmedOps.length === 0 ? (
-                <p className="text-sm text-muted-foreground">
-                  {t('orders.noConfirmedOperations')}
-                </p>
-              ) : (
-                <div className="rounded-md border p-3 max-h-48 overflow-y-auto space-y-1">
-                  {[...confirmedOps]
-                    .sort((a, b) => (a.planned_date_earliest || '').localeCompare(b.planned_date_earliest || ''))
-                    .map((op) => (
-                    <label
-                      key={op.id}
-                      className="flex items-center gap-2 text-sm py-0.5"
-                    >
-                      <input
-                        type="checkbox"
-                        checked={selectedOpIds.includes(op.id)}
-                        onChange={() => handleOpToggle(op.id)}
-                        className="rounded border-input"
-                      />
-                      <span>
-                        #{op.id} — {op.short_description || t('orders.noDescription')}
-                      </span>
-                      {op.planned_date_earliest && (
-                        <span className="text-muted-foreground">
-                          ({op.planned_date_earliest}
-                          {op.planned_date_latest
-                            ? ` – ${op.planned_date_latest}`
-                            : ""}
-                          )
-                        </span>
-                      )}
-                    </label>
-                  ))}
-                </div>
-              )}
-            </div>
-
-            {/* Estimated route km */}
-            <div className="space-y-2">
-              <Label htmlFor="routeKm">{t('orders.estimatedRouteKm')} *</Label>
-              <Input
-                id="routeKm"
-                type="number"
-                min={1}
-                value={estimatedRouteKm}
-                onChange={(e) => setEstimatedRouteKm(e.target.value)}
-                required
-              />
-            </div>
-
-            {/* Create-mode map preview */}
-            {(() => {
-              const createMapOps = createModeOpDetails
-                .filter((op) => op.route_coordinates && op.route_coordinates.length > 0)
-                .map((op) => ({ id: op.id, route_coordinates: op.route_coordinates! }));
-              const createStartSite = landingSites.find((s) => s.id === Number(startSiteId));
-              const createEndSite = landingSites.find((s) => s.id === Number(endSiteId));
-              const hasMapData = createMapOps.length > 0 || createStartSite || createEndSite;
-              if (!hasMapData) return null;
-              return (
-                <div className="space-y-2">
-                  <Label>{t('orders.routeMap')}</Label>
-                  <div key={`${selectedOpIds.join(',')}-${startSiteId}-${endSiteId}`}>
-                    <OrderMap
-                      operations={createMapOps}
-                      startLandingSite={
-                        createStartSite
-                          ? { name: createStartSite.name, lat: createStartSite.latitude, lng: createStartSite.longitude }
-                          : null
-                      }
-                      endLandingSite={
-                        createEndSite
-                          ? { name: createEndSite.name, lat: createEndSite.latitude, lng: createEndSite.longitude }
-                          : null
-                      }
-                    />
-                  </div>
-                </div>
-              );
-            })()}
-
-            <div className="flex gap-3 pt-2">
-              <Button type="submit" disabled={createMutation.isPending}>
-                {createMutation.isPending
-                  ? t('orders.creating')
-                  : t('orders.createOrder')}
-              </Button>
-              <Button
-                type="button"
-                variant="outline"
-                onClick={() => navigate("/orders")}
-              >
-                {t('orders.cancel')}
-              </Button>
-            </div>
-          </form>
-        </div>
-      ) : (
-        /* ═══════════ DETAIL MODE ═══════════ */
-        <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-          {/* Left: Order details */}
-          <div className="space-y-6">
-            <div className="rounded-md bg-surface-container-low p-6 space-y-4">
-              <h2 className="text-lg font-semibold">{t('orders.orderDetails')}</h2>
-
-              <div className="grid grid-cols-2 gap-4">
-                <div>
-                  <Label className="text-muted-foreground">
-                    {t('orders.plannedStartLabel')}
-                  </Label>
-                  <p className="text-sm font-medium">
-                    {order
-                      ? new Date(
-                          order.planned_start_datetime
-                        ).toLocaleString("pl-PL")
-                      : "—"}
-                  </p>
-                </div>
-                <div>
-                  <Label className="text-muted-foreground">
-                    {t('orders.plannedEndLabel')}
-                  </Label>
-                  <p className="text-sm font-medium">
-                    {order
-                      ? new Date(
-                          order.planned_end_datetime
-                        ).toLocaleString("pl-PL")
-                      : "—"}
-                  </p>
-                </div>
-              </div>
-
-              <div className="grid grid-cols-2 gap-4">
-                <div>
-                  <Label className="text-muted-foreground">{t('orders.helicopter')}</Label>
-                  <p className="text-sm font-medium">
-                    {order?.helicopter_registration ?? "—"}
-                  </p>
-                </div>
-                <div>
-                  <Label className="text-muted-foreground">
-                    {t('orders.estimatedRoute')}
-                  </Label>
-                  <p className="text-sm font-medium">
-                    {order?.estimated_route_km != null
-                      ? `${order.estimated_route_km} km`
-                      : "—"}
-                  </p>
-                </div>
-              </div>
-
-              <div>
-                <Label className="text-muted-foreground">{t('orders.crewWeightLabel')}</Label>
-                <p className="text-sm font-medium">
-                  {order?.crew_weight != null ? `${order.crew_weight} kg` : "—"}
-                </p>
-              </div>
-
-              <div className="grid grid-cols-2 gap-4">
-                <div>
-                  <Label className="text-muted-foreground">
-                    {t('orders.startLandingSiteLabel')}
-                  </Label>
-                  <p className="text-sm font-medium">
-                    {order?.start_landing_site_name ?? "—"}
-                  </p>
-                </div>
-                <div>
-                  <Label className="text-muted-foreground">
-                    {t('orders.endLandingSiteLabel')}
-                  </Label>
-                  <p className="text-sm font-medium">
-                    {order?.end_landing_site_name ?? "—"}
-                  </p>
-                </div>
-              </div>
-            </div>
-
-            {/* Actual datetimes — editable by pilot when status=4 */}
-            {isPilot && currentStatus === 4 && (
-              <div className="rounded-md bg-surface-container-low p-6 space-y-4">
-                <h2 className="text-lg font-semibold">
-                  {t('orders.actualTimes')}
-                </h2>
-                <div className="grid grid-cols-2 gap-4">
-                  <div className="space-y-2">
-                    <Label htmlFor="actualStart">{t('orders.actualStart')}</Label>
-                    <Input
-                      id="actualStart"
-                      type="datetime-local"
-                      value={actualStart}
-                      onChange={(e) => setActualStart(e.target.value)}
-                    />
-                  </div>
-                  <div className="space-y-2">
-                    <Label htmlFor="actualEnd">{t('orders.actualEnd')}</Label>
-                    <Input
-                      id="actualEnd"
-                      type="datetime-local"
-                      value={actualEnd}
-                      onChange={(e) => setActualEnd(e.target.value)}
-                    />
-                  </div>
-                </div>
-                <Button
-                  onClick={() => updateMutation.mutate()}
-                  disabled={updateMutation.isPending}
-                >
-                  {updateMutation.isPending
-                    ? t('orders.savingTimes')
-                    : t('orders.saveTimes')}
-                </Button>
-              </div>
-            )}
-
-            {/* Read-only actual datetimes when not editable */}
-            {(!isPilot || currentStatus !== 4) &&
-              (order?.actual_start_datetime || order?.actual_end_datetime) && (
-                <div className="rounded-md bg-surface-container-low p-6 space-y-4">
-                  <h2 className="text-lg font-semibold">
-                    {t('orders.actualTimes')}
-                  </h2>
-                  <div className="grid grid-cols-2 gap-4">
-                    <div>
-                      <Label className="text-muted-foreground">
-                        {t('orders.actualStart')}
-                      </Label>
-                      <p className="text-sm font-medium">
-                        {order.actual_start_datetime
-                          ? new Date(
-                              order.actual_start_datetime
-                            ).toLocaleString("pl-PL")
-                          : "—"}
-                      </p>
-                    </div>
-                    <div>
-                      <Label className="text-muted-foreground">
-                        {t('orders.actualEnd')}
-                      </Label>
-                      <p className="text-sm font-medium">
-                        {order.actual_end_datetime
-                          ? new Date(
-                              order.actual_end_datetime
-                            ).toLocaleString("pl-PL")
-                          : "—"}
-                      </p>
-                    </div>
-                  </div>
-                </div>
-              )}
-
-            {/* Crew members list */}
-            <div className="rounded-md bg-surface-container-low p-6">
-              <h2 className="text-lg font-semibold mb-3">{t('orders.crewSection')}</h2>
-              {order && order.crew_members.length > 0 ? (
-                <ul className="space-y-1">
-                  {order.crew_members.map((cm) => (
-                    <li key={cm.id} className="text-sm">
-                      {cm.name}{" "}
-                      <span className="text-muted-foreground">
-                        ({cm.weight} kg)
-                      </span>
-                    </li>
-                  ))}
-                </ul>
-              ) : (
-                <p className="text-sm text-muted-foreground">{t('orders.noCrew')}</p>
-              )}
-            </div>
-
-            {/* Operations list */}
-            <div className="rounded-md bg-surface-container-low p-6">
-              <h2 className="text-lg font-semibold mb-3">{t('orders.operationsSection')}</h2>
-              {order && order.operations.length > 0 ? (
-                <ul className="space-y-1">
-                  {order.operations.map((op) => (
-                    <li key={op.id} className="text-sm">
-                      #{op.id} — {op.short_description ?? t('orders.noDescription')}{" "}
-                      <Badge variant="outline" className="ml-1 text-xs">
-                        {t(`operations.status${op.status}`, { defaultValue: `Status ${op.status}` })}
-                      </Badge>
-                    </li>
-                  ))}
-                </ul>
-              ) : (
-                <p className="text-sm text-muted-foreground">
-                  {t('orders.noLinkedOperations')}
-                </p>
-              )}
-            </div>
-          </div>
-
-          {/* Right: Map */}
-          <div className="space-y-6">
-            {(mapOperations.length > 0 || startSiteForMap || endSiteForMap) && (
-              <div className="rounded-md bg-surface-container-low p-6">
-                <h2 className="text-lg font-semibold mb-3">
-                  {t('orders.routeMap')}
-                </h2>
-                <OrderMap
-                  operations={mapOperations}
-                  startLandingSite={
-                    startSiteForMap
-                      ? {
-                          name: startSiteForMap.name,
-                          lat: startSiteForMap.latitude,
-                          lng: startSiteForMap.longitude,
-                        }
-                      : null
-                  }
-                  endLandingSite={
-                    endSiteForMap
-                      ? {
-                          name: endSiteForMap.name,
-                          lat: endSiteForMap.latitude,
-                          lng: endSiteForMap.longitude,
-                        }
-                      : null
-                  }
-                />
-              </div>
-            )}
-          </div>
-        </div>
-      )}
+        <OrderCreateForm
+          userFullName={user ? `${user.first_name} ${user.last_name}` : ""}
+          userEmail={user?.email ?? ""}
+          plannedStart={plannedStart}
+          plannedEnd={plannedEnd}
+          helicopterId={helicopterId}
+          selectedCrewIds={selectedCrewIds}
+          startSiteId={startSiteId}
+          endSiteId={endSiteId}
+          selectedOpIds={selectedOpIds}
+          estimatedRouteKm={estimatedRouteKm}
+          helicopters={helicopters}
+          allCrew={allCrew}
+          landingSites={landingSites}
+          confirmedOps={confirmedOps}
+          createModeOpDetails={createModeOpDetails}
+          onPlannedStartChange={setPlannedStart}
+          onPlannedEndChange={setPlannedEnd}
+          onHelicopterIdChange={setHelicopterId}
+          onCrewToggle={handleCrewToggle}
+          onStartSiteIdChange={setStartSiteId}
+          onEndSiteIdChange={setEndSiteId}
+          onOpToggle={handleOpToggle}
+          onEstimatedRouteKmChange={setEstimatedRouteKm}
+          onSubmit={handleSubmit}
+          onCancel={() => navigate("/orders")}
+          isCreating={createMutation.isPending}
+        />
+      ) : order ? (
+        <>
+          <OrderStatusActions
+            status={currentStatus}
+            isPilot={isPilot}
+            isSupervisor={isSupervisor}
+            hasActualTimes={!!order.actual_start_datetime && !!order.actual_end_datetime}
+            onSubmit={() => submitMutation.mutate()}
+            onAccept={() => acceptMutation.mutate()}
+            onReject={() => setShowRejectDialog(true)}
+            onCompletePartial={() => completePartialMutation.mutate()}
+            onCompleteFull={() => completeFullMutation.mutate()}
+            onNotCompleted={() => setShowNotCompletedDialog(true)}
+            submitPending={submitMutation.isPending}
+            acceptPending={acceptMutation.isPending}
+            rejectPending={rejectMutation.isPending}
+            completePartialPending={completePartialMutation.isPending}
+            completeFullPending={completeFullMutation.isPending}
+            notCompletedPending={notCompletedMutation.isPending}
+          />
+          <OrderDetailView
+            order={order}
+            isPilot={isPilot}
+            currentStatus={currentStatus}
+            actualStart={actualStart}
+            actualEnd={actualEnd}
+            onActualStartChange={setActualStart}
+            onActualEndChange={setActualEnd}
+            onSaveActualTimes={() => updateMutation.mutate()}
+            savingActualTimes={updateMutation.isPending}
+            mapOperations={mapOperations}
+            startSiteForMap={
+              startSiteForMap
+                ? { name: startSiteForMap.name, lat: startSiteForMap.latitude, lng: startSiteForMap.longitude }
+                : null
+            }
+            endSiteForMap={
+              endSiteForMap
+                ? { name: endSiteForMap.name, lat: endSiteForMap.latitude, lng: endSiteForMap.longitude }
+                : null
+            }
+          />
+        </>
+      ) : null}
 
       {/* Reject confirm dialog */}
       <Dialog open={showRejectDialog} onOpenChange={() => setShowRejectDialog(false)}>
@@ -1110,110 +645,6 @@ export function OrderFormPage() {
           </DialogFooter>
         </DialogContent>
       </Dialog>
-    </div>
-  );
-}
-
-// ── OrderStatusActions sub-component ────────────────────────────────
-
-function OrderStatusActions({
-  status,
-  isPilot,
-  isSupervisor,
-  hasActualTimes,
-  onSubmit,
-  onAccept,
-  onReject,
-  onCompletePartial,
-  onCompleteFull,
-  onNotCompleted,
-  submitPending,
-  acceptPending,
-  rejectPending,
-  completePartialPending,
-  completeFullPending,
-  notCompletedPending,
-}: {
-  status: number;
-  isPilot: boolean;
-  isSupervisor: boolean;
-  hasActualTimes: boolean;
-  onSubmit: () => void;
-  onAccept: () => void;
-  onReject: () => void;
-  onCompletePartial: () => void;
-  onCompleteFull: () => void;
-  onNotCompleted: () => void;
-  submitPending: boolean;
-  acceptPending: boolean;
-  rejectPending: boolean;
-  completePartialPending: boolean;
-  completeFullPending: boolean;
-  notCompletedPending: boolean;
-}) {
-  const { t } = useTranslation();
-  const showSubmit = isPilot && status === 1;
-  const showSupervisorActions = isSupervisor && status === 2;
-  const showSettlement = isPilot && status === 4;
-
-  if (!showSubmit && !showSupervisorActions && !showSettlement) return null;
-
-  return (
-    <div className="mb-6 flex flex-wrap gap-3">
-      {showSubmit && (
-        <Button onClick={onSubmit} disabled={submitPending}>
-          {submitPending ? t('orders.submitting') : t('orders.submitForAcceptance')}
-        </Button>
-      )}
-      {showSupervisorActions && (
-        <>
-          <Button
-            onClick={onAccept}
-            disabled={acceptPending}
-            className="bg-green-600 hover:bg-green-700"
-          >
-            <CheckCircle className="mr-2 h-4 w-4" />
-            {acceptPending ? t('orders.accepting') : t('orders.accept')}
-          </Button>
-          <Button
-            variant="destructive"
-            onClick={onReject}
-            disabled={rejectPending}
-          >
-            <XCircle className="mr-2 h-4 w-4" />
-            {rejectPending ? t('orders.rejecting') : t('orders.reject')}
-          </Button>
-        </>
-      )}
-      {showSettlement && (
-        <>
-          <Button
-            onClick={onCompletePartial}
-            disabled={completePartialPending || !hasActualTimes}
-            className="bg-orange-500 hover:bg-orange-600"
-          >
-            {completePartialPending
-              ? t('orders.processing')
-              : t('orders.completePartial')}
-          </Button>
-          <Button
-            onClick={onCompleteFull}
-            disabled={completeFullPending || !hasActualTimes}
-            className="bg-green-600 hover:bg-green-700"
-          >
-            {completeFullPending
-              ? t('orders.processing')
-              : t('orders.completeFull')}
-          </Button>
-          <Button
-            variant="destructive"
-            onClick={onNotCompleted}
-            disabled={notCompletedPending}
-          >
-            {notCompletedPending ? t('orders.processing') : t('orders.notCompleted')}
-          </Button>
-        </>
-      )}
     </div>
   );
 }
